added parallelization options for some functions

+++ R/helper_functions.R
@@ -12,6 +12,7 @@
 #' @param location character; path to directory with files/directories you want to manipulate.
 #' @param move.from character; path to directory you want to move files/directories from.
 #' @param move.to character; path to directory you want to move files/directories to.
+#' @param n.cores integer; if 1, will run serially, otherwise will extract in parallel utilzing up to the number of cores specified. Default is 1.
 #' @seealso \code{\link{system}}, \code{\link{generate.full.commands}}, \code{\link{generate.tool.commands}}
 
 #' @rdname helper.functions
@@ -24,7 +25,7 @@ show.tool.help <- function(tool) {
 #' @rdname helper.functions
 #' @export
 
-tgz.directories <- function(location, match.pattern = NULL) {
+tgz.directories <- function(location, match.pattern = NULL, n.cores = 1) {
   require(magrittr)
   require(stringr)
   if (is.null(match.pattern)) {
@@ -34,8 +35,22 @@ tgz.directories <- function(location, match.pattern = NULL) {
       str_subset(match.pattern)
   }
   if (length(target.dirs) > 1) {
-    for (target.dir in target.dirs) {
-      tar(tarfile = paste0(target.dir, ".tgz"), files = target.dir, compression = "gzip")
+    if (n.cores == 1) {
+      for (target.dir in target.dirs) {
+        tar(tarfile = paste0(target.dir, ".tgz"), files = target.dir, compression = "gzip")
+      }
+    } else {
+      require(doParallel)
+      require(foreach)
+      cl <- parallel::makeCluster(n.cores, type = "FORK")
+      doParallel::registerDoParallel(cl, n.cores)
+      par.loop <- foreach::foreach(target.dir = target.dirs) %dopar% {
+        tar(tarfile = paste0(target.dir, ".tgz"), files = target.dir, compression = "gzip")
+      } %>% try(silent = T)
+      parallel::stopCluster(cl)
+      if ("try-error" %in% class(par.loop)) {
+        cat(par.loop, sep = "\n")
+      }
     }
   } else if (length(target.dir) == 0) {
     rlang::inform("No directories detected, nothing done.")
@@ -47,7 +62,7 @@ tgz.directories <- function(location, match.pattern = NULL) {
 #' @rdname helper.functions
 #' @export
 
-remove.directories <- function(location, match.pattern = NULL) {
+remove.directories <- function(location, match.pattern = NULL, n.cores = 1) {
   require(magrittr)
   require(stringr)
   if (is.null(match.pattern)) {
@@ -59,9 +74,24 @@ remove.directories <- function(location, match.pattern = NULL) {
   if (length(target.dirs) == 0) {
     rlang::inform("No directories detected, nothing done.")
   } else {
-    for (target.dir in target.dirs) {
-      file.remove(list.files(path = target.dir, full.names = T, recursive = T))
-      file.remove(target.dir)
+    if (n.cores == 1) {
+      for (target.dir in target.dirs) {
+        file.remove(list.files(path = target.dir, full.names = T, recursive = T))
+        file.remove(target.dir)
+      }
+    } else {
+      require(doParallel)
+      require(foreach)
+      cl <- parallel::makeCluster(n.cores, type = "FORK")
+      doParallel::registerDoParallel(cl, n.cores)
+      par.loop <- foreach::foreach(target.dir = target.dirs) %dopar% {
+        file.remove(list.files(path = target.dir, full.names = T, recursive = T))
+        file.remove(target.dir)
+      } %>% try(silent = T)
+      parallel::stopCluster(cl)
+      if ("try-error" %in% class(par.loop)) {
+        cat(par.loop, sep = "\n")
+      }
     }
   }
 }
@@ -69,38 +99,83 @@ remove.directories <- function(location, match.pattern = NULL) {
 #' @rdname helper.functions
 #' @export
 
-gzip.files <- function(location, match.pattern = NULL) {
+gzip.files <- function(location, match.pattern = NULL, n.cores = 1) {
   files <- list.files(path = location, pattern = match.pattern, full.names = T)
-  for (file in files) {
-    cmd <- paste("gzip -v", file)
-    system(cmd)
+  if (n.cores == 1) {
+    for (file in files) {
+      cmd <- paste("gzip -v", file)
+      system(cmd)
+    }
+  } else {
+    require(doParallel)
+    require(foreach)
+    cl <- parallel::makeCluster(n.cores, type = "FORK")
+    doParallel::registerDoParallel(cl, n.cores)
+    par.loop <- foreach::foreach(file = files) %dopar% {
+      cmd <- paste("gzip -v", file)
+      system(cmd)
+    } %>% try(silent = T)
+    parallel::stopCluster(cl)
+    if ("try-error" %in% class(par.loop)) {
+      cat(par.loop, sep = "\n")
+    }
   }
 }
 
 #' @rdname helper.functions
 #' @export
 
-gunzip.files <- function(location, match.pattern = NULL) {
+gunzip.files <- function(location, match.pattern = NULL, n.cores = 1) {
   files <- list.files(path = location, pattern = match.pattern, full.names = T)
-  for (file in files) {
-    cmd <- paste("gunzip -v", file)
-    system(cmd)
+  if (n.cores == 1) {
+    for (file in files) {
+      cmd <- paste("gunzip -v", file)
+      system(cmd)
+    }
+  } else {
+    require(doParallel)
+    require(foreach)
+    cl <- parallel::makeCluster(n.cores, type = "FORK")
+    doParallel::registerDoParallel(cl, n.cores)
+    par.loop <- foreach::foreach(file = files) %dopar% {
+      cmd <- paste("gunzip -v", file)
+      system(cmd)
+    } %>% try(silent = T)
+    parallel::stopCluster(cl)
+    if ("try-error" %in% class(par.loop)) {
+      cat(par.loop, sep = "\n")
+    }
   }
 }
 
 #' @rdname helper.functions
 #' @export
 
-move.files <- function(move.from, move.to, match.pattern = NULL) {
+move.files <- function(move.from, move.to, match.pattern = NULL, n.cores = 1) {
   require(magrittr)
   if (is.null(match.pattern)) {
     copy.res <- file.copy(from = move.from, to = move.to) %>% try()
     if (!{"try-error" %in% class(copy.res)}) { file.remove(move.from) }
   } else {
     files <- list.files(path = move.from, pattern = match.pattern, full.names = T)
-    for (file in files) {
-      copy.res <- file.copy(from = file, to = move.to) %>% try()
-      if (!{"try-error" %in% class(copy.res)}) { file.remove(file) }
+    if (n.cores == 1) {
+      for (file in files) {
+        copy.res <- file.copy(from = file, to = move.to) %>% try()
+        if (!{"try-error" %in% class(copy.res)}) { file.remove(file) }
+      }
+    } else {
+      require(doParallel)
+      require(foreach)
+      cl <- parallel::makeCluster(n.cores, type = "FORK")
+      doParallel::registerDoParallel(cl, n.cores)
+      par.loop <- foreach::foreach(file = files) %dopar% {
+        copy.res <- file.copy(from = file, to = move.to) %>% try()
+        if (!{"try-error" %in% class(copy.res)}) { file.remove(file) }
+      } %>% try(silent = T)
+      parallel::stopCluster(cl)
+      if ("try-error" %in% class(par.loop)) {
+        cat(par.loop, sep = "\n")
+      }
     }
   }
 }
